+++ command/job_inspect.go
@@ -94,7 +94,7 @@ func (c *JobInspectCommand) Run(args []string) int {
 	}
 
 	// If args not specified but output format is specified, format and output the jobs data list
-	if len(args) == 0 && json || len(tmpl) > 0 {
+	if len(args) == 0 && (json || len(tmpl) > 0) {
 		jobs, _, err := client.Jobs().List(nil)
 		if err != nil {
 			c.Ui.Error(fmt.Sprintf("Error querying jobs: %v", err))