+++ command/scaling_policy_list.go
@@ -42,6 +42,9 @@ Policy Info Options:
   -type
     Filter scaling policies by type.
 
+  -verbose
+    Display full information.
+
   -json
     Output the scaling policy in its JSON format.
 
@@ -59,10 +62,11 @@ func (s *ScalingPolicyListCommand) Synopsis() string {
 func (s *ScalingPolicyListCommand) AutocompleteFlags() complete.Flags {
 	return mergeAutocompleteFlags(s.Meta.AutocompleteFlags(FlagSetClient),
 		complete.Flags{
-			"-job":  complete.PredictNothing,
-			"-type": complete.PredictNothing,
-			"-json": complete.PredictNothing,
-			"-t":    complete.PredictAnything,
+			"-verbose": complete.PredictNothing,
+			"-job":     complete.PredictNothing,
+			"-type":    complete.PredictNothing,
+			"-json":    complete.PredictNothing,
+			"-t":       complete.PredictAnything,
 		})
 }
 
@@ -71,11 +75,12 @@ func (s *ScalingPolicyListCommand) Name() string { return "scaling policy list" 
 
 // Run satisfies the cli.Command Run function.
 func (s *ScalingPolicyListCommand) Run(args []string) int {
-	var json bool
+	var json, verbose bool
 	var tmpl, policyType, job string
 
 	flags := s.Meta.FlagSet(s.Name(), FlagSetClient)
 	flags.Usage = func() { s.Ui.Output(s.Help()) }
+	flags.BoolVar(&verbose, "verbose", false, "")
 	flags.BoolVar(&json, "json", false, "")
 	flags.StringVar(&tmpl, "t", "", "")
 	flags.StringVar(&policyType, "type", "", "")
@@ -87,6 +92,13 @@ func (s *ScalingPolicyListCommand) Run(args []string) int {
 	if args = flags.Args(); len(args) > 0 {
 		s.Ui.Error("This command takes no arguments")
 		s.Ui.Error(commandErrorText(s))
+		return 1
+	}
+
+	// Truncate the id unless full length is requested
+	length := shortId
+	if verbose {
+		length = fullId
 	}
 
 	// Get the HTTP client.
@@ -111,11 +123,6 @@ func (s *ScalingPolicyListCommand) Run(args []string) int {
 		return 1
 	}
 
-	if len(policies) == 0 {
-		s.Ui.Output("No policies found")
-		return 0
-	}
-
 	if json || len(tmpl) > 0 {
 		out, err := Format(json, tmpl, policies)
 		if err != nil {
@@ -126,23 +133,33 @@ func (s *ScalingPolicyListCommand) Run(args []string) int {
 		return 0
 	}
 
+	output := formatScalingPolicies(policies, length)
+	s.Ui.Output(output)
+	return 0
+}
+
+func formatScalingPolicies(stubs []*api.ScalingPolicyListStub, uuidLength int) string {
+	if len(stubs) == 0 {
+		return "No policies found"
+	}
+
 	// Create the output table header.
-	output := []string{"ID|Enabled|Type|Target"}
+	policies := []string{"ID|Enabled|Type|Target"}
 
 	// Sort the list of policies based on their target.
-	sortedPolicies := scalingPolicyStubList{policies: policies}
+	sortedPolicies := scalingPolicyStubList{policies: stubs}
 	sort.Sort(sortedPolicies)
 
 	// Iterate the policies and add to the output.
 	for _, policy := range sortedPolicies.policies {
-		output = append(output, fmt.Sprintf(
+		policies = append(policies, fmt.Sprintf(
 			"%s|%v|%s|%s",
-			policy.ID, policy.Enabled, policy.Type, formatScalingPolicyTarget(policy.Target)))
+			limit(policy.ID, uuidLength),
+			policy.Enabled,
+			policy.Type,
+			formatScalingPolicyTarget(policy.Target)))
 	}
-
-	// Output.
-	s.Ui.Output(formatList(output))
-	return 0
+	return formatList(policies)
 }
 
 // scalingPolicyStubList is a wrapper around []*api.ScalingPolicyListStub that